Netlist updates:
- Removed trampolines (OUTLOGIC, INPLOGIC and friends). 
- Started using doxygen comment and documentation style. Added doxygen
files to documentation folder. 
- Refactored code triggered by doxygen output. 
- Moved internal and support classes into namespace detail. 
- Use an anordered map in parser. 
- -Wconversion fixes - All done now. 
- Fixed -Wold-style-cast warnings in netlist code. 
- Added iterators to pstring. 
- Moved two macros, added more RAII and improved exceptions. Fixed some
bugs in parser code. 
- Fixed a number of bugs in parser code and exception handling. 
[Couriersud]

+++ src/lib/netlist/plib/palloc.h
@@ -17,7 +17,6 @@
 #include "pstring.h"
 
 namespace plib {
-
 //============================================================
 //  exception base
 //============================================================
@@ -204,7 +203,7 @@ private:
 		char *data;
 	};
 
-	int new_block();
+	size_t new_block();
 
 	struct info
 	{
@@ -213,14 +212,14 @@ private:
 	};
 
 public:
-	mempool(int min_alloc, int min_align);
+	mempool(size_t min_alloc, size_t min_align);
 	~mempool();
 
 	void *alloc(size_t size);
 	void free(void *ptr);
 
-	int m_min_alloc;
-	int m_min_align;
+	size_t m_min_alloc;
+	size_t m_min_align;
 
 	std::vector<block> m_blocks;
 };
